Use dtype

+++ tensorflow/models/rnn/translate/translate.py
@@ -164,7 +164,8 @@ def create_model(session, forward_only):
       FLAGS.learning_rate,
       FLAGS.learning_rate_decay_factor,
       forward_only=forward_only,
-      num_samples=-1)
+      num_samples=-1,
+      dtype=dtype)
 
   ckpt = tf.train.get_checkpoint_state(FLAGS.train_dir)
   if ckpt and tf.train.checkpoint_exists(ckpt.model_checkpoint_path):
